reorg api types
