+++ api/v1alpha1/ext_proc_types.go
@@ -1,5 +1,73 @@
 package v1alpha1
 
+import (
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+)
+
+// ExtProcProvider defines the configuration for an ExtProc provider.
+type ExtProcProvider struct {
+	// GrpcService is the GRPC service that will handle the processing.
+	// +required
+	GrpcService *ExtGrpcService `json:"grpcService"`
+
+	// FailOpen determines if requests are allowed when the ext proc service is unavailable.
+	// Defaults to true, meaning requests are allowed upstream even if the ext proc service is unavailable.
+	// +optional
+	// +kubebuilder:default=true
+	FailOpen bool `json:"failOpen,omitempty"`
+
+	// ProcessingMode defines how the filter should interact with the request/response streams.
+	// +optional
+	ProcessingMode *ProcessingMode `json:"processingMode,omitempty"`
+
+	// MessageTimeout is the timeout for each message sent to the external processing server.
+	// +optional
+	// +kubebuilder:validation:XValidation:rule="matches(self, '^([0-9]{1,5}(h|m|s|ms)){1,4}$')",message="invalid timeout value"
+	// +kubebuilder:validation:XValidation:rule="duration(self) >= duration('1ms')",message="timeout must be at least 1ms."
+	MessageTimeout *metav1.Duration `json:"messageTimeout,omitempty"`
+
+	// MaxMessageTimeout specifies the upper bound of override_message_timeout that may be sent from the external processing server.
+	// The default value 0, which effectively disables the override_message_timeout API.
+	// +optional
+	// +kubebuilder:validation:XValidation:rule="matches(self, '^([0-9]{1,5}(h|m|s|ms)){1,4}$')",message="invalid timeout value"
+	// +kubebuilder:validation:XValidation:rule="duration(self) >= duration('1ms')",message="timeout must be at least 1ms."
+	MaxMessageTimeout *metav1.Duration `json:"maxMessageTimeout,omitempty"`
+
+	// StatPrefix is an optional prefix to include when emitting stats from the extproc filter,
+	// enabling different instances of the filter to have unique stats.
+	// +optional
+	// +kubebuilder:validation:MinLength=1
+	StatPrefix *string `json:"statPrefix,omitempty"`
+
+	// RouteCacheAction describes the route cache action to be taken when an
+	// external processor response is received in response to request headers.
+	// The default behavior is to only clear the route cache when an external processing
+	// response has the clear_route_cache field set.
+	// +optional
+	// +kubebuilder:validation:Enum=Default;Clear;Retain
+	// +kubebuilder:default=Default
+	RouteCacheAction ExtProcRouteCacheAction `json:"routeCacheAction,omitempty"`
+
+	// MetadataOptions allows configuring metadata namespaces to forwarded or received from the external
+	// processing server.
+	// +optional
+	MetadataOptions *MetadataOptions `json:"metadataOptions,omitempty"`
+}
+
+type ExtProcRouteCacheAction string
+
+const (
+	// RouteCacheActionDefault is the default behavior, which clears the route cache only
+	// when the clear_route_cache field is set in an external processor response.
+	RouteCacheActionDefault ExtProcRouteCacheAction = "Default"
+	// RouteCacheActionClear always clears the route cache irrespective of the
+	// clear_route_cache field in the external processor response.
+	RouteCacheActionClear ExtProcRouteCacheAction = "Clear"
+	// RouteCacheActionRetain never clears the route cache irrespective of the
+	// clear_route_cache field in the external processor response.
+	RouteCacheActionRetain ExtProcRouteCacheAction = "Retain"
+)
+
 // ExtProcPolicy defines the configuration for the Envoy External Processing filter.
 //
 // +kubebuilder:validation:ExactlyOneOf=extensionRef;disable
